refactor(constants): tmp 제거
파일명의 대소문자 변화는 git에서 인식하지 못합니다.
