+++ app/src/@core/guards/UnAuthGuard.tsx
@@ -1,5 +1,5 @@
 import { useAuth } from '@core/guards/hooks/useAuth';
-import { ROUTES } from '@shared/constants/(tmp)routes';
+import { ROUTES } from '@shared/constants/routes';
 import { Navigate, Outlet } from 'react-router-dom';
 
 export const UnAuthGuard = () => {
