+++ app/src/@shared/components/DashboardContentView/Rank/ProjectRankListItem.tsx
@@ -1,5 +1,5 @@
 import { ProjectRank } from '@shared/__generated__/graphql';
-import { ROUTES } from '@shared/constants/(tmp)routes';
+import { ROUTES } from '@shared/constants/routes';
 import { RankListItem } from './RankListItem';
 
 type ProjectRankListItemProps = {
