+++ app/src/@shared/components/DashboardContentView/Rank/UserRankListItem.tsx
@@ -1,6 +1,6 @@
 import { UserRank } from '@shared/__generated__/graphql';
-import { ROUTES } from '@shared/constants/(tmp)routes';
 import { ALT } from '@shared/constants/accessibility';
+import { ROUTES } from '@shared/constants/routes';
 import { Avatar } from '@shared/ui-kit';
 import { RankListItem } from './RankListItem';
