+++ app/src/Home/constants/quotes.ts
