+++ app/src/Leaderboard/components/Leaderboard/LeaderboardListItem.tsx
@@ -1,8 +1,8 @@
 import { useTheme } from '@emotion/react';
 import styled from '@emotion/styled';
 import { UserRank } from '@shared/__generated__/graphql';
-import { ROUTES } from '@shared/constants/(tmp)routes';
 import { ALT } from '@shared/constants/accessibility';
+import { ROUTES } from '@shared/constants/routes';
 import {
   Avatar,
   Body1MediumText,
