+++ app/src/Profile/dashboard-contents/Eval/EvalLogSearchLink.tsx
@@ -1,6 +1,6 @@
 import { UserProfileContext } from '@/Profile/contexts/UserProfileContext';
 import { DashboardContent } from '@shared/components/DashboardContent';
-import { ROUTES } from '@shared/constants/(tmp)routes';
+import { ROUTES } from '@shared/constants/routes';
 import { Center } from '@shared/ui-kit';
 import { CustomLink } from '@shared/ui-kit-styled/CustomLink';
 import { useContext } from 'react';
