+++ app/src/Profile/dashboard-contents/General/TeamInfo/TeamInfoTable.tsx
@@ -1,6 +1,6 @@
 import styled from '@emotion/styled';
 import { UserTeam } from '@shared/__generated__/graphql';
-import { ROUTES } from '@shared/constants/(tmp)routes';
+import { ROUTES } from '@shared/constants/routes';
 import { PrimaryMediumText, Text } from '@shared/ui-kit';
 import { getDateDiffStringWithTeamStatus } from '@shared/utils/getDateDiffStringWithTeamStatus';
 import { rgba } from 'emotion-rgba';
