+++ app/src/Project/dashboard-contents/PastEvaluationLink.tsx
@@ -1,5 +1,5 @@
 import { DashboardContent } from '@shared/components/DashboardContent';
-import { ROUTES } from '@shared/constants/(tmp)routes';
+import { ROUTES } from '@shared/constants/routes';
 import { CustomLink } from '@shared/ui-kit-styled/CustomLink';
 import { useContext } from 'react';
 import { ProjectNameContext } from '../contexts/ProjectNameContext';
